Remove the refresh flag from bulk indexing
We have a conflict between the production and staging versions
of elasticsearch. Staging is running v1.6.0 and production is running
v1.4.4.

There is an undocumented breaking change between the 1.4 version and
1.6: the bulk indexer does not accept a "refresh" flag in the metadata
per document. It does, however accept a ?refresh argument as a query
param, so if we ever need this, we can add that behavior to the bulk
sender Send() method, which calls DoCommand().

We have never used this feature (it is expensive, we don't need it,
we've always passed refresh=false).